Projects/Personal_Travel_Assistant(PTA)/Personal_Travel_Assistant(PTA)/requirements.txt
fastapi==0.95.2            # framework for building apis using python 
uvicorn[standard]==0.22.0  # ASGI server to run fastapi
motor==3.1.1               # Async motor to connect mongodb driver
pydantic==1.10.12          # used for data validations and to make ordered data  
httpx==0.24.1              # it was used to call dummy apis used in this project
python-dotenv==1.0.0       # it loads .env file
